Use TypeScript
