+++ package.json
@@ -20,7 +20,7 @@
     "clean": "rm -r dist",
     "lint": "eslint . && prettier . --check && npm run rule-syntax-check",
     "bundle": "./build.sh",
-    "rule-syntax-check": "node scripts/validate-json-rules.js",
+    "rule-syntax-check": "ts-node scripts/validate-json-rules.ts",
     "watch": "npm run prepublish && chokidar \"lib\" \"addon\" \"rules/autoconsent\" \"rules/filterlist.txt\" --ignore 'lib/filterlist-engine.ts' -c \"npm run prepublish\"",
     "create-rule": "node rules/create-rule.mjs",
     "test": "playwright test",
